wip: more permissions

+++ sql/revert/permissions.sql
@@ -4,15 +4,24 @@ BEGIN;
 
 revoke all on schema _api from public;
 
-revoke all on schema api from anonymous;
-revoke all on all routines in schema _api from anonymous;
-revoke all on all routines in schema  api from anonymous;
+do $$
+begin
+  if exists(select 1 from pg_roles where rolname = 'god') then
+    drop role god;
+  end if;
+  if exists(select 1 from pg_roles where rolname = 'authenticated') then
+    revoke all on all routines in schema api from authenticated;
+    revoke all on all tables in schema api from authenticated;
+    drop role authenticated;
+  end if;
 
-revoke all on all routines in schema api from authenticated;
-revoke all on all tables in schema api from authenticated;
+  if exists(select 1 from pg_roles where rolname = 'anonymous') then
+    revoke all on schema api from anonymous;
+    revoke all on all routines in schema _api from anonymous;
+    revoke all on all routines in schema  api from anonymous;
+    drop role anonymous;
+  end if;
 
-drop role god;
-drop role authenticated;
-drop role anonymous;
+end $$;
 
 COMMIT;
